caqti index

Library caqti

This library exposes the following toplevel modules:

Caqti_common Prerequisites.

Caqti_common_priv Internal: Prerequisites.

Caqti_connect Connection functor and backend registration.

Caqti_connect_sig Signature for establishing database connections.

Caqti_connection Internal: Connection Utilities for Drivers

Caqti_connection_sig Signature of connection handles.

Caqti_driver_info Information about a database, its driver, and its query language.

Caqti_driver_lib Internal: Library for Drivers

Caqti_driver_sig Internal: Signature for driver implementation.

Caqti_error Error descriptors.

Caqti_heap Internal: Min-heap.

Caqti_mult Row multiplicity.

Caqti_pool Internal: Resource pool.

Caqti_pool_sig Resource pool signature.

Caqti_query Query specification.

Caqti_request Request specification.

Caqti_response_sig Signature of a response from a database.

Caqti_sql SQL utility functions.

Caqti_sql_io SQL IO utility functions.

Caqti_stream A stream with monadic concurrency and error handling.

Caqti_type Type descriptors for fields and tuples.

Caqti_type_sig Signatures for Caqti_type.

Library caqti.blocking

This library exposes the following toplevel modules:

Caqti_blocking Blocking API based on the Unix module.

Caqti_blocking_sql_io