caqti index

Library caqti

This library exposes the following toplevel modules:

Caqti_connect_sig Signatures providing functions for establishing database connections.

Caqti_connection_sig Signature of connection handles.

Caqti_driver_info Information about a database, its driver, and its query language.

Caqti_error Error descriptors.

Caqti_mult Row multiplicity.

Caqti_pool_config Pool configuration.

Caqti_pool_sig Resource pool signature.

Caqti_query Query specification.

Caqti_request Request specification.

Caqti_response_sig Signature of a response from a database.

Caqti_stream_sig Concurrent stream signature.

Caqti_switch_sig Switch implementation used where not available.

Caqti_type Type descriptors for fields and tuples.

Caqti_type_sig Signatures for Caqti_type.

Library caqti.blocking

This library implements the blocking (non-)concurrency using the unix library. Real concurrency support is provided by separate packages.

The entry point of this library is the module: Caqti_blocking.

Platform Libraries for Internal Use

The platform libraries are only meant for use in implementing drivers and concurrency support. These APIs are unstable, i.e. they can change between minor version and without prior deprecation notices.

Library caqti.platform

The entry point of this library is the module: Caqti_platform.

Library caqti.platform.unix

The entry point of this library is the module: Caqti_platform_unix.